Make basic enemy. Restructures player and enemy... creates "Entity" class
This allows for enemies and players to be stored inside of a vector list within the room. Now we can have them interacting on a separate level than the ground etc.
Also updated the about.txt for ppl who want to edit levels!
Dang it! There's no design for the game yet!!!! Dang it!!!!