Makes gameloop.h handle the window dimensions (also fixes camera disorientation)
So yeah, basically the camera works like normal now, except for the fact that cutscenes don't resize, the actual game window expands, keeping the 1 pixel the same size eh.

(TODO) maybe later the game will be able to scale its size by 2 or 3 if the player wants hd textures, that way they have both the option of hd or large large screen space.

So then, for cutscenes, automatically just scaling it to what the size should be should be the best option???? Or a function to rescale it via an animation